Исправления: 1. Исправлен regexp шаблон и алгоритм для нахождения URL в строках. 2. Исправлено отклбчение пользовательских стилей для строк в дереве событий программы (вложенные строки отображались с пользовательскими настройками даже при включении стиля системы). 3. Убрана светло-серая рамка вокруг панели меню (некрасиво смотрится при темных стилях оформления системы).

+++ trunk/REXLoader/rexwindow.cpp
@@ -1424,7 +1424,7 @@ void REXWindow::syncTaskData()
         downtimeId = mdl->mapToSource(downtimeId);
         speedAvgId = mdl->mapToSource(speedAvgId);
         int downtime = model->data(downtimeId,100).toInt();
-        int speedAvg = downtime ? (model->data(speedAvgId,100).toLongLong()*downtime + speed)/(downtime+1) : 0;
+        qint64 speedAvg = downtime ? (model->data(speedAvgId,100).toLongLong()*downtime + speed)/(downtime+1) : 0;
         ++downtime;
         delete(mdl);
